Explicitly separate submit handlers
We're passing a submit handler that gets called by an internal handler,
so I've decided to use some names to make it more obvious.

The consumer of our CredentialsForm just passes a "submitHandler"
property, they don't have to think about internal vs external...
but inside the component, it's a different story.

The internalSubmitHandler prepares data and passes it to the
externalSubmitHandler.

Hopefully this makes stuff more obvious.